Move a file `LinkData.tx` from `\modules\wallet` to `\modules\vote`
