+++ src/index.ts
@@ -34,7 +34,7 @@ export { ProposalType, ProposalData } from "./modules/vote/ProposalData";
 export { VoterCard, BallotData } from "./modules/vote/BallotData";
 export { Encrypt } from "./modules/vote/Encrypt";
 export { EncryptionKey } from "./modules/vote/EncryptionKey";
-export { LinkDataWithProposalFee, LinkDataWithProposalData, LinkDataWithVoteData } from "./modules/wallet/LinkData";
+export { LinkDataWithProposalFee, LinkDataWithProposalData, LinkDataWithVoteData } from "./modules/vote/LinkData";
 
 export { LockType, Lock, Unlock } from "./modules/script/Lock";
 export { OP, isOpcode, isConditional, isPayload } from "./modules/script/Opcodes";
