+++ src/modules/vote/LinkData.ts
