+++ src/modules/vote/ProposalFeeData.ts
@@ -15,7 +15,7 @@
 import { PublicKey } from "../common/KeyPair";
 import { VarInt } from "../utils/VarInt";
 import { Utils } from "../utils/Utils";
-import { LinkDataWithProposalFee } from "../wallet/LinkData";
+import { LinkDataWithProposalFee } from "./LinkData";
 
 import JSBI from "jsbi";
 import { SmartBuffer } from "smart-buffer";
